Move all stubs into separate directory, refactor Makefile 1 lab for this and refactor stubs for this
