+++ tests/stubs/plugin/greeting/bad/greeting_bad_init.c
@@ -1,3 +1,5 @@
+// tests/stubs/plugin/greeting_bad_init.c
+
 #include "../../src/include/master.h"
 #include <stdio.h>
 
@@ -5,15 +7,14 @@ static Hook last_executor_start_hook = NULL;
 
 void custom_executor_start_hook(void);
 
-void init(void) {
-    last_executor_start_hook = executor_start_hook;
-    executor_start_hook = custom_executor_start_hook;
-    printf("greeting initialized\n");
-}
-
 const char *name(void) {
     return "greeting";
 }
+
+void fini(void) {
+    printf("greeting finished\n");
+}
+
 
 void custom_executor_start_hook(void) {
     printf("Hello, world!\n");
